requirements.txt
python-dotenv
pandas 
streamlit
langchain
langchain-groq